// src/api/generated/schemas/patchProfileBodyBody.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * ClothingFit API
 * ClothingFit 백엔드 API 문서
 * OpenAPI spec version: 1.0.0
 */

export type PatchProfileBodyBody = {
  height: number;
  weight: number;
  chest?: number;
  waist?: number;
  hip?: number;
  shoulder?: number;
  head?: number;
  footSize?: number;
};